+++ src/python/pants/engine/target.py
@@ -14,13 +14,11 @@ import textwrap
 import zlib
 from abc import ABC, ABCMeta, abstractmethod
 from collections import deque
-from collections.abc import Callable, Iterable, Iterator, KeysView, Mapping, Sequence
+from collections.abc import Callable, Iterable, Iterator, Mapping, Sequence
 from dataclasses import dataclass
 from enum import Enum
-from operator import attrgetter
 from pathlib import PurePath
 from typing import (
-    AbstractSet,
     Any,
     ClassVar,
     Generic,
@@ -31,7 +29,6 @@ from typing import (
     get_type_hints,
 )
 
-from pants.base.deprecated import warn_or_error
 from pants.engine.addresses import Address, Addresses, UnparsedAddressInputs, assert_single_address
 from pants.engine.collection import Collection
 from pants.engine.engine_aware import EngineAwareParameter
@@ -63,6 +60,7 @@ from pants.engine.internals.native_engine import (
 from pants.engine.internals.native_engine import SourcesField as SourcesField
 from pants.engine.internals.native_engine import StringField as StringField
 from pants.engine.internals.native_engine import StringSequenceField as StringSequenceField
+from pants.engine.internals.native_engine import Target as Target  # noqa: F401
 from pants.engine.internals.native_engine import TriBoolField as TriBoolField  # noqa: F401
 from pants.engine.internals.target_adaptor import SourceBlock, SourceBlocks  # noqa: F401
 from pants.engine.rules import rule
@@ -163,437 +161,7 @@ class FieldDefaults:
 _F = TypeVar("_F", bound=Field)
 
 
-@dataclass(frozen=True)
-class Target:
-    """A Target represents an addressable set of metadata.
-
-    Set the `help` class property with a description, which will be used in `./pants help`. For the
-    best rendering, use soft wrapping (e.g. implicit string concatenation) within paragraphs, but
-    hard wrapping (`\n`) to separate distinct paragraphs and/or lists.
-    """
-
-    # Subclasses must define these
-    alias: ClassVar[str]
-    core_fields: ClassVar[tuple[type[Field], ...]]
-    help: ClassVar[str | Callable[[], str]]
-
-    removal_version: ClassVar[str | None] = None
-    removal_hint: ClassVar[str | None] = None
-
-    deprecated_alias: ClassVar[str | None] = None
-    deprecated_alias_removal_version: ClassVar[str | None] = None
-
-    # These get calculated in the constructor
-    address: Address
-    field_values: FrozenDict[type[Field], Field]
-    residence_dir: str
-    name_explicitly_set: bool
-    description_of_origin: str
-    origin_sources_blocks: FrozenDict[str, SourceBlocks]
-
-    @final
-    def __init__(
-        self,
-        unhydrated_values: Mapping[str, Any],
-        address: Address,
-        # NB: `union_membership` is only optional to facilitate tests. In production, we should
-        # always provide this parameter. This should be safe to do because production code should
-        # rarely directly instantiate Targets and should instead use the engine to request them.
-        union_membership: UnionMembership | None = None,
-        *,
-        name_explicitly_set: bool = True,
-        residence_dir: str | None = None,
-        ignore_unrecognized_fields: bool = False,
-        description_of_origin: str | None = None,
-        origin_sources_blocks: FrozenDict[str, SourceBlocks] = FrozenDict(),
-    ) -> None:
-        """Create a target.
-
-        :param unhydrated_values: A mapping of field aliases to their raw values. Any left off
-            fields will either use their default or error if required=True.
-        :param address: How to uniquely identify this target.
-        :param union_membership: Used to determine plugin fields. This must be set in production!
-        :param residence_dir: Where this target "lives". If unspecified, will be the `spec_path`
-            of the `address`, i.e. where the target was either explicitly defined or where its
-            target generator was explicitly defined. Target generators can, however, set this to
-            the directory where the generated target provides metadata for. For example, a
-            file-based target like `python_source` should set this to the parent directory of
-            its file. A file-less target like `go_third_party_package` should keep the default of
-            `address.spec_path`. This field impacts how command line specs work, so that globs
-            like `dir:` know whether to match the target or not.
-        :param ignore_unrecognized_fields: Don't error if fields are not recognized. This is only
-            intended for when Pants is bootstrapping itself.
-        :param description_of_origin: Where this target was declared, such as a path to BUILD file
-            and line number.
-        """
-        if self.removal_version and not address.is_generated_target:
-            if not self.removal_hint:
-                raise ValueError(
-                    f"You specified `removal_version` for {self.__class__}, but not "
-                    "the class property `removal_hint`."
-                )
-            warn_or_error(
-                self.removal_version,
-                entity=f"the {repr(self.alias)} target type",
-                hint=f"Using the `{self.alias}` target type for {address}. {self.removal_hint}",
-            )
-
-        if origin_sources_blocks:
-            _validate_origin_sources_blocks(origin_sources_blocks)
-
-        object.__setattr__(
-            self, "residence_dir", residence_dir if residence_dir is not None else address.spec_path
-        )
-        object.__setattr__(self, "address", address)
-        object.__setattr__(
-            self, "description_of_origin", description_of_origin or self.residence_dir
-        )
-        object.__setattr__(self, "origin_sources_blocks", origin_sources_blocks)
-        object.__setattr__(self, "name_explicitly_set", name_explicitly_set)
-        try:
-            object.__setattr__(
-                self,
-                "field_values",
-                self._calculate_field_values(
-                    unhydrated_values,
-                    address,
-                    union_membership,
-                    ignore_unrecognized_fields=ignore_unrecognized_fields,
-                ),
-            )
-
-            self.validate()
-        except Exception as e:
-            raise InvalidTargetException(
-                str(e), description_of_origin=self.description_of_origin
-            ) from e
-
-    @final
-    def _calculate_field_values(
-        self,
-        unhydrated_values: Mapping[str, Any],
-        address: Address,
-        # See `__init__`.
-        union_membership: UnionMembership | None,
-        *,
-        ignore_unrecognized_fields: bool,
-    ) -> FrozenDict[type[Field], Field]:
-        all_field_types = self.class_field_types(union_membership)
-        field_values = {}
-        aliases_to_field_types = self._get_field_aliases_to_field_types(all_field_types)
-
-        for alias, value in unhydrated_values.items():
-            if alias not in aliases_to_field_types:
-                if ignore_unrecognized_fields:
-                    continue
-                valid_aliases = set(aliases_to_field_types.keys())
-                if isinstance(self, TargetGenerator):
-                    # Even though moved_fields don't live on the target generator, they are valid
-                    # for users to specify. It's intentional that these are only used for
-                    # `InvalidFieldException` and are not stored as normal fields with
-                    # `aliases_to_field_types`.
-                    for field_type in self.moved_fields:
-                        valid_aliases.add(field_type.alias)
-                        if field_type.deprecated_alias is not None:
-                            valid_aliases.add(field_type.deprecated_alias)
-                raise InvalidFieldException(
-                    f"Unrecognized field `{alias}={value}` in target {address}. Valid fields for "
-                    f"the target type `{self.alias}`: {sorted(valid_aliases)}.",
-                )
-            field_type = aliases_to_field_types[alias]
-            field_values[field_type] = field_type(value, address)
-
-        # For undefined fields, mark the raw value as missing.
-        for field_type in all_field_types:
-            if field_type in field_values:
-                continue
-            field_values[field_type] = field_type(NO_VALUE, address)
-        return FrozenDict(
-            sorted(
-                field_values.items(),
-                key=lambda field_type_to_val_pair: field_type_to_val_pair[0].alias,
-            )
-        )
-
-    @final
-    @classmethod
-    def _get_field_aliases_to_field_types(
-        cls, field_types: Iterable[type[Field]]
-    ) -> dict[str, type[Field]]:
-        aliases_to_field_types = {}
-        for field_type in field_types:
-            aliases_to_field_types[field_type.alias] = field_type
-            if field_type.deprecated_alias is not None:
-                aliases_to_field_types[field_type.deprecated_alias] = field_type
-        return aliases_to_field_types
-
-    @final
-    @property
-    def field_types(self) -> KeysView[type[Field]]:
-        return self.field_values.keys()
-
-    @distinct_union_type_per_subclass
-    class PluginField:
-        pass
-
-    def __repr__(self) -> str:
-        fields = ", ".join(str(field) for field in self.field_values.values())
-        return (
-            f"{self.__class__}("
-            f"address={self.address}, "
-            f"alias={self.alias!r}, "
-            f"residence_dir={self.residence_dir!r}, "
-            f"origin={self.description_of_origin}, "
-            f"{fields})"
-        )
-
-    def __str__(self) -> str:
-        fields = ", ".join(str(field) for field in self.field_values.values())
-        address = f'address="{self.address}"{", " if fields else ""}'
-        return f"{self.alias}({address}{fields})"
-
-    def __hash__(self) -> int:
-        return hash((self.__class__, self.address, self.residence_dir, self.field_values))
-
-    def __eq__(self, other: Target | Any) -> bool:
-        if not isinstance(other, Target):
-            return NotImplemented
-        return (self.__class__, self.address, self.residence_dir, self.field_values) == (
-            other.__class__,
-            other.address,
-            other.residence_dir,
-            other.field_values,
-        )
-
-    def __lt__(self, other: Any) -> bool:
-        if not isinstance(other, Target):
-            return NotImplemented
-        return self.address < other.address
-
-    def __gt__(self, other: Any) -> bool:
-        if not isinstance(other, Target):
-            return NotImplemented
-        return self.address > other.address
-
-    @classmethod
-    @memoized_method
-    def _find_plugin_fields(cls, union_membership: UnionMembership) -> tuple[type[Field], ...]:
-        result: set[type[Field]] = set()
-        classes = [cls]
-        while classes:
-            cls = classes.pop()
-            classes.extend(cls.__bases__)
-            if issubclass(cls, Target):
-                result.update(cast("set[type[Field]]", union_membership.get(cls.PluginField)))
-
-        return tuple(sorted(result, key=attrgetter("alias")))
-
-    @final
-    @classmethod
-    def _find_registered_field_subclass(
-        cls, requested_field: type[_F], *, registered_fields: Iterable[type[Field]]
-    ) -> type[_F] | None:
-        """Check if the Target has registered a subclass of the requested Field.
-
-        This is necessary to allow targets to override the functionality of common fields. For
-        example, you could subclass `Tags` to define `CustomTags` with a different default. At the
-        same time, we still want to be able to call `tgt.get(Tags)`, in addition to
-        `tgt.get(CustomTags)`.
-        """
-        subclass = next(
-            (
-                registered_field
-                for registered_field in registered_fields
-                if issubclass(registered_field, requested_field)
-            ),
-            None,
-        )
-        return subclass
-
-    @final
-    def _maybe_get(self, field: type[_F]) -> _F | None:
-        result = self.field_values.get(field, None)
-        if result is not None:
-            return cast(_F, result)
-        field_subclass = self._find_registered_field_subclass(
-            field, registered_fields=self.field_values.keys()
-        )
-        if field_subclass is not None:
-            return cast(_F, self.field_values[field_subclass])
-        return None
-
-    @final
-    def __getitem__(self, field: type[_F]) -> _F:
-        """Get the requested `Field` instance belonging to this target.
-
-        If the `Field` is not registered on this `Target` type, this method will raise a
-        `KeyError`. To avoid this, you should first call `tgt.has_field()` or `tgt.has_fields()`
-        to ensure that the field is registered, or, alternatively, use `Target.get()`.
-
-        See the docstring for `Target.get()` for how this method handles subclasses of the
-        requested Field and for tips on how to use the returned value.
-        """
-        result = self._maybe_get(field)
-        if result is not None:
-            return result
-        raise KeyError(
-            f"The target `{self}` does not have a field `{field.__name__}`. Before calling "
-            f"`my_tgt[{field.__name__}]`, call `my_tgt.has_field({field.__name__})` to "
-            f"filter out any irrelevant Targets or call `my_tgt.get({field.__name__})` to use the "
-            f"default Field value."
-        )
-
-    @final
-    def get(self, field: type[_F], *, default_raw_value: Any | None = None) -> _F:
-        """Get the requested `Field` instance belonging to this target.
-
-        This will return an instance of the requested field type, e.g. an instance of
-        `InterpreterConstraints`, `SourcesField`, `EntryPoint`, etc. Usually, you will want to
-        grab the `Field`'s inner value, e.g. `tgt.get(Compatibility).value`. (For async fields like
-        `SourcesField`, you may need to hydrate the value.).
-
-        This works with subclasses of `Field`. For example, if you subclass `Tags`
-        to define a custom subclass `CustomTags`, both `tgt.get(Tags)` and
-        `tgt.get(CustomTags)` will return the same `CustomTags` instance.
-
-        If the `Field` is not registered on this `Target` type, this will return an instance of
-        the requested Field by using `default_raw_value` to create the instance. Alternatively,
-        first call `tgt.has_field()` or `tgt.has_fields()` to ensure that the field is registered,
-        or, alternatively, use indexing (e.g. `tgt[Compatibility]`) to raise a KeyError when the
-        field is not registered.
-        """
-        result = self._maybe_get(field)
-        if result is not None:
-            return result
-        return field(default_raw_value, self.address)
-
-    @final
-    @classmethod
-    def _has_fields(
-        cls, fields: Iterable[type[Field]], *, registered_fields: AbstractSet[type[Field]]
-    ) -> bool:
-        unrecognized_fields = [field for field in fields if field not in registered_fields]
-        if not unrecognized_fields:
-            return True
-        for unrecognized_field in unrecognized_fields:
-            maybe_subclass = cls._find_registered_field_subclass(
-                unrecognized_field, registered_fields=registered_fields
-            )
-            if maybe_subclass is None:
-                return False
-        return True
-
-    @final
-    def has_field(self, field: type[Field]) -> bool:
-        """Check that this target has registered the requested field.
-
-        This works with subclasses of `Field`. For example, if you subclass `Tags` to define a
-        custom subclass `CustomTags`, both `tgt.has_field(Tags)` and
-        `python_tgt.has_field(CustomTags)` will return True.
-        """
-        return self.has_fields([field])
-
-    @final
-    def has_fields(self, fields: Iterable[type[Field]]) -> bool:
-        """Check that this target has registered all of the requested fields.
-
-        This works with subclasses of `Field`. For example, if you subclass `Tags` to define a
-        custom subclass `CustomTags`, both `tgt.has_fields([Tags])` and
-        `python_tgt.has_fields([CustomTags])` will return True.
-        """
-        return self._has_fields(fields, registered_fields=self.field_values.keys())
-
-    @final
-    @classmethod
-    @memoized_method
-    def class_field_types(
-        cls, union_membership: UnionMembership | None
-    ) -> FrozenOrderedSet[type[Field]]:
-        """Return all registered Fields belonging to this target type.
-
-        You can also use the instance property `tgt.field_types` to avoid having to pass the
-        parameter UnionMembership.
-        """
-        if union_membership is None:
-            return FrozenOrderedSet(cls.core_fields)
-        else:
-            return FrozenOrderedSet((*cls.core_fields, *cls._find_plugin_fields(union_membership)))
-
-    @final
-    @classmethod
-    def class_has_field(cls, field: type[Field], union_membership: UnionMembership) -> bool:
-        """Behaves like `Target.has_field()`, but works as a classmethod rather than an instance
-        method."""
-        return cls.class_has_fields([field], union_membership)
-
-    @final
-    @classmethod
-    def class_has_fields(
-        cls, fields: Iterable[type[Field]], union_membership: UnionMembership
-    ) -> bool:
-        """Behaves like `Target.has_fields()`, but works as a classmethod rather than an instance
-        method."""
-        return cls._has_fields(fields, registered_fields=cls.class_field_types(union_membership))
-
-    @final
-    @classmethod
-    def class_get_field(cls, field: type[_F], union_membership: UnionMembership) -> type[_F]:
-        """Get the requested Field type registered with this target type.
-
-        This will error if the field is not registered, so you should call Target.class_has_field()
-        first.
-        """
-        class_fields = cls.class_field_types(union_membership)
-        result = next(
-            (
-                registered_field
-                for registered_field in class_fields
-                if issubclass(registered_field, field)
-            ),
-            None,
-        )
-        if result is None:
-            raise KeyError(
-                f"The target type `{cls.alias}` does not have a field `{field.__name__}`. Before "
-                f"calling `TargetType.class_get_field({field.__name__})`, call "
-                f"`TargetType.class_has_field({field.__name__})`."
-            )
-        return result
-
-    @classmethod
-    def register_plugin_field(cls, field: type[Field]) -> UnionRule:
-        """Register a new field on the target type.
-
-        In the `rules()` register.py entry-point, include
-        `MyTarget.register_plugin_field(NewField)`. This will register `NewField` as a first-class
-        citizen. Plugins can use this new field like any other.
-        """
-        return UnionRule(cls.PluginField, field)
-
-    def validate(self) -> None:
-        """Validate the target, such as checking for mutually exclusive fields.
-
-        N.B.: The validation should only be of properties intrinsic to the associated files in any
-        context. If the validation only makes sense for certain goals acting on targets; those
-        validations should be done in the associated rules.
-        """
-
-
-def _validate_origin_sources_blocks(origin_sources_blocks: FrozenDict[str, SourceBlocks]) -> None:
-    if not isinstance(origin_sources_blocks, FrozenDict):
-        raise ValueError(
-            f"Expected `origin_sources_blocks` to be of type `FrozenDict`, got {type(origin_sources_blocks)=} {origin_sources_blocks=}"
-        )
-    for blocks in origin_sources_blocks.values():
-        if not isinstance(blocks, SourceBlocks):
-            raise ValueError(
-                f"Expected `origin_sources_blocks` to be a `FrozenDict` with values of type `SourceBlocks`, got values of {type(blocks)=} {blocks=}"
-            )
-        for block in blocks:
-            if not isinstance(block, SourceBlock):
-                raise ValueError(
-                    f"Expected `origin_sources_blocks` to be a `FrozenDict` with values of type `SourceBlocks`, got values of {type(blocks)=} {blocks=}"
-                )
+# Target is imported from pants.engine.internals.native_engine above.
 
 
 @dataclass(frozen=True)
