+++ src/python/pants/engine/target_test.py
@@ -1,11 +1,10 @@
 # Copyright 2020 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-import re
 import string
 from collections import namedtuple
 from collections.abc import Iterable, Sequence
-from dataclasses import FrozenInstanceError, dataclass
+from dataclasses import dataclass
 from enum import Enum
 from typing import Any, ClassVar, cast
 
@@ -48,7 +47,6 @@ from pants.engine.target import (
     Target,
     TupleSequenceField,
     ValidNumbers,
-    _validate_origin_sources_blocks,
     generate_file_based_overrides_field_help_message,
     get_shard,
     parse_shard_spec,
@@ -141,8 +139,8 @@ def test_field_and_target_eq() -> None:
     assert hash(tgt) != hash(other_tgt)
 
     # Ensure the target is frozen.
-    with pytest.raises(FrozenInstanceError):
-        tgt.address = addr  # type: ignore[misc]
+    with pytest.raises(AttributeError):
+        tgt.address = addr
 
     # Ensure that subclasses are not equal.
     class SubclassField(FortranVersion):
@@ -1576,19 +1574,22 @@ def test_generate_file_based_overrides_field_help_message() -> None:
 
 
 def test_validate_origin_sources_blocks():
-    with pytest.raises(ValueError, match=re.compile("^Expected .*`FrozenDict`, got .*list")):
-        _validate_origin_sources_blocks([SourceBlock(start=0, end=1)])
-    with pytest.raises(
-        ValueError,
-        match=re.compile(r"^Expected .*`SourceBlocks`, got .*SourceBlock"),
-    ):
-        _validate_origin_sources_blocks(FrozenDict([("file.txt", SourceBlock(start=0, end=1))]))
-    with pytest.raises(
-        ValueError,
-        match=re.compile(r"^Expected .*`SourceBlocks`, got .*tuple"),
-    ):
-        _validate_origin_sources_blocks(FrozenDict([("file.txt", ((0, 1),))]))
+    """Validation is now in Rust Target.__new__. Test through Target construction."""
+    addr = Address("", target_name="t")
 
-    _validate_origin_sources_blocks(
-        FrozenDict([("file.txt", SourceBlocks([SourceBlock(start=0, end=1)]))])
+    # Valid case
+    FortranTarget(
+        {},
+        addr,
+        origin_sources_blocks=FrozenDict(
+            [("file.txt", SourceBlocks([SourceBlock(start=0, end=1)]))]
+        ),
     )
+
+    # Invalid: values must be SourceBlocks
+    with pytest.raises(Exception, match="SourceBlocks"):
+        FortranTarget(
+            {},
+            addr,
+            origin_sources_blocks=FrozenDict([("file.txt", SourceBlock(start=0, end=1))]),
+        )
